Extracted the defaults into their own file.
