+++ packages/log/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @idearium/log
 
+## Unreleased
+
+### Changed
+
+-   Extracted the defaults into their own file.
+
 ## v1.0.0 - 2021-05-06
 
 -   First version of the package.
